Rename soc.h to libsteel.h
